Add code review suggestions
