+++ workspaces/ballerina/ballerina-core/src/rpc-types/bi-diagram/index.ts
@@ -110,8 +110,7 @@ import {
     RecordsInWorkspaceMentions,
     BuildMode,
     DevantMetadata,
-    GeneratedClientSaveResponse,
-    MigrateRequest
+    GeneratedClientSaveResponse
 } from "./interfaces";
 
 export interface BIDiagramAPI {
@@ -128,7 +127,6 @@ export interface BIDiagramAPI {
     getNodeTemplate: (params: BINodeTemplateRequest) => Promise<BINodeTemplateResponse>;
     getAiSuggestions: (params: BIAiSuggestionsRequest) => Promise<BIAiSuggestionsResponse>;
     createProject: (params: ProjectRequest) => void;
-    migrateProject: (params: MigrateRequest) => void;
     getWorkspaces: () => Promise<WorkspacesResponse>;
     getProjectStructure: () => Promise<ProjectStructureResponse>;
     getProjectComponents: () => Promise<ProjectComponentsResponse>;
