+++ workspaces/ballerina/ballerina-core/src/rpc-types/migrate-integration/index.ts
@@ -17,7 +17,7 @@
  */
 
 import { ImportIntegrationResponse } from "../../interfaces/extended-lang-client";
-import { GetMigrationToolsResponse, ImportIntegrationRPCRequest, MigrationToolPullRequest, OpenMigrationReportRequest, SaveMigrationReportRequest } from "./interfaces";
+import { GetMigrationToolsResponse, ImportIntegrationRPCRequest, MigrateRequest, MigrationToolPullRequest, OpenMigrationReportRequest, SaveMigrationReportRequest } from "./interfaces";
 
 export interface MigrateIntegrationAPI {
     getMigrationTools: () => Promise<GetMigrationToolsResponse>;
@@ -25,4 +25,5 @@ export interface MigrateIntegrationAPI {
     importIntegration: (params: ImportIntegrationRPCRequest) => Promise<ImportIntegrationResponse>;
     openMigrationReport: (params: OpenMigrationReportRequest) => void;
     saveMigrationReport: (params: SaveMigrationReportRequest) => void;
+    migrateProject: (params: MigrateRequest) => void;
 }
