+++ workspaces/ballerina/ballerina-extension/src/rpc-managers/migrate-integration/rpc-manager.ts
@@ -23,11 +23,12 @@ import {
     ImportIntegrationResponse,
     ImportIntegrationRPCRequest,
     MigrateIntegrationAPI,
+    MigrateRequest,
     OpenMigrationReportRequest,
     SaveMigrationReportRequest
 } from "@wso2/ballerina-core";
 import { StateMachine } from "../../stateMachine";
-import { getUsername, sanitizeName } from "../../utils/bi";
+import { createBIProjectFromMigration, getUsername, sanitizeName } from "../../utils/bi";
 import { pullMigrationTool } from "../../utils/migrate-integration";
 import { MigrationReportWebview } from "../../views/migration-report/webview";
 
@@ -71,7 +72,7 @@ export class MigrateIntegrationRpcManager implements MigrateIntegrationAPI {
 
     async saveMigrationReport(params: SaveMigrationReportRequest): Promise<void> {
         const vscode = await import('vscode');
-        
+
         // Show save dialog
         const saveUri = await vscode.window.showSaveDialog({
             defaultUri: vscode.Uri.file(params.defaultFileName),
@@ -80,11 +81,15 @@ export class MigrateIntegrationRpcManager implements MigrateIntegrationAPI {
                 'All files': ['*']
             }
         });
-        
+
         if (saveUri) {
             // Write the report content to the selected file
             await vscode.workspace.fs.writeFile(saveUri, Buffer.from(params.reportContent, 'utf8'));
             vscode.window.showInformationMessage(`Migration report saved to ${saveUri.fsPath}`);
         }
+    }
+
+    async migrateProject(params: MigrateRequest): Promise<void> {
+        createBIProjectFromMigration(params);
     }
 }
