+++ workspaces/ballerina/ballerina-rpc-client/src/rpc-clients/bi-diagram/rpc-client.ts
@@ -73,7 +73,6 @@ import {
     JsonToTypeRequest,
     JsonToTypeResponse,
     LinePosition,
-    MigrateRequest,
     ModelFromCodeRequest,
     OpenAPIClientDeleteRequest,
     OpenAPIClientDeleteResponse,
@@ -166,7 +165,6 @@ import {
     getVisibleVariableTypes,
     getWorkspaces,
     handleReadmeContent,
-    migrateProject,
     openAIChat,
     openConfigToml,
     openReadme,
@@ -245,10 +243,6 @@ export class BiDiagramRpcClient implements BIDiagramAPI {
 
     createProject(params: ProjectRequest): void {
         return this._messenger.sendNotification(createProject, HOST_EXTENSION, params);
-    }
-
-    migrateProject(params: MigrateRequest): void {
-        return this._messenger.sendNotification(migrateProject, HOST_EXTENSION, params);
     }
 
     getWorkspaces(): Promise<WorkspacesResponse> {
